cleanup home component data
