+++ lib/planet-list.ts
@@ -1,0 +1,51 @@
+import { solarSystemPlanet } from './types';
+
+const planetsList: solarSystemPlanet[] = [
+  {
+    planetName: 'Mercury',
+    visibleRings: false,
+    planetWidth: 'w-16',
+  },
+  {
+    planetName: 'Venus',
+    visibleRings: false,
+    planetWidth: 'w-20',
+  },
+  {
+    planetName: 'Earth',
+    visibleRings: false,
+    planetWidth: 'w-24',
+  },
+  {
+    planetName: 'Mars',
+    visibleRings: false,
+    planetWidth: 'w-20',
+  },
+  {
+    planetName: 'Jupiter',
+    visibleRings: false,
+    planetWidth: 'w-44',
+  },
+  {
+    planetName: 'Saturn',
+    visibleRings: true,
+    planetWidth: 'w-36',
+  },
+  {
+    planetName: 'Uranus',
+    visibleRings: true,
+    planetWidth: 'w-32',
+  },
+  {
+    planetName: 'Neptune',
+    visibleRings: false,
+    planetWidth: 'w-28',
+  },
+  {
+    planetName: 'Pluto',
+    visibleRings: false,
+    planetWidth: 'w-20',
+  },
+];
+
+export default planetsList;
